Fix to use app router + edit home page
