--- public/json/albums.json
@@ -1,12 +1,0 @@
-[
-  {
-    "title": "The Lie",
-    "imageUrl": "/media/the-lie-demo.webp",
-    "releaseDate": "July 12 2024"
-  },
-  {
-    "title": "Killing Twins Demo",
-    "imageUrl": "/media/demo-art.webp",
-    "releaseDate": "2024"
-  }
-]
