--- public/json/events.json
@@ -1,1 +1,0 @@
-[]
