+++ src/app/layout.tsx
@@ -1,13 +1,9 @@
-import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import '@/styles/globals.css';
+import "@/styles/globals.css";
+import Navbar from "@/components/navbar"; // Import directly without dynamic import
+import Script from "next/script";
 
 const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
 
 export default function RootLayout({
   children,
@@ -16,7 +12,15 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <head>
+        <Script src="http://localhost:8097"></Script>
+      </head>
+      <body className={inter.className}>
+        <main className="min-h-dvh bg-black text-white">
+          <Navbar />
+          {children}
+        </main>
+      </body>
     </html>
   );
 }
