+++ src/json/events.json
@@ -1,0 +1,1 @@
+[]
