+++ src/json/media.json
@@ -1,5 +1,10 @@
 [
   {
+    "src": "/media/the-lie-demo.webp",
+    "alt": "The Lie",
+    "size": 600
+  },
+  {
     "src": "/media/demo-art.webp",
     "alt": "Killing Twins Demo",
     "size": 600
