--- src/pages/_app.tsx
@@ -1,7 +1,0 @@
-// These styles apply to every route in the application
-import '@/styles/globals.css'
-import type { AppProps } from 'next/app'
- 
-export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
-}
